Health bars for player and damage/health for NPC and player now work

+++ Assets/Scripts/NPCDefault.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "NPCDefault", menuName = "Scriptable Objects/NPCDefault")]
 public class NPCDefault : ScriptableObject
 {
-    public float maxHealth;
+    public int maxHealth;
     public float angerIncrement, wrongDrinkIncrement, maxAngerCap, minAngerCap;
